+++ src/index.ts
@@ -1,10 +1,10 @@
-import { xior } from './xior';
+import { Xior } from './xior';
 
 export * from './xior';
 export * from './types';
 // @ts-ignore
 export * from 'xior/utils';
 
-const xiorObj = Object.assign(xior.create(), { create: xior.create, VERSION: xior.VERSION });
+const xiorObj = Object.assign(Xior.create(), { create: Xior.create, VERSION: Xior.VERSION });
 
 export default xiorObj;