+++ tests/src/tests/index.test.ts
@@ -3,7 +3,7 @@ import { readFile, writeFile } from 'node:fs/promises';
 import { before, after, describe, it } from 'node:test';
 // @ts-ignore
 import stringify from 'qs/lib/stringify';
-import { isXiorError, merge, xior } from 'xior';
+import { isXiorError, merge, Xior } from 'xior';
 import errorRetryPlugin from 'xior/plugins/error-retry';
 
 import { readChunks, startServer } from './server';
@@ -21,12 +21,12 @@ after(async () => {
 
 describe('xior tests', () => {
   describe('should work with methods GET/POST/PUT/PATCH/DELETE/HEAD/OPTIONS', () => {
-    const xiorInstance = xior.create({ baseURL });
+    const xiorInstance = Xior.create({ baseURL });
 
-    it('xior.VERSION should be same with package.json', async () => {
+    it('Xior.VERSION should be same with package.json', async () => {
       const content = await readFile('../package.json', 'utf-8');
       const pkg = JSON.parse(content);
-      assert.strictEqual(xior.VERSION, pkg.version);
+      assert.strictEqual(Xior.VERSION, pkg.version);
     });
 
     it('GET should work', async () => {
@@ -141,7 +141,7 @@ describe('xior tests', () => {
   });
 
   describe('Options `paramsSerializer` and `encodeURI` tests', () => {
-    const xiorInstance = xior.create({ baseURL, withCredentials: true });
+    const xiorInstance = Xior.create({ baseURL, withCredentials: true });
 
     it('default encode should work', async () => {
       const { data, request } = await xiorInstance.get<{
@@ -167,7 +167,7 @@ describe('xior tests', () => {
     });
 
     it('Use `qs.stringify` as custom encode function should work', async () => {
-      const xiorInstance = xior.create({
+      const xiorInstance = Xior.create({
         baseURL,
         paramsSerializer: (params: Record<string, any>) => stringify(params, {}),
       });
@@ -187,9 +187,9 @@ describe('xior tests', () => {
     });
   });
 
-  describe('`xior.interceptors` tests', () => {
-    it('`xior.interceptors.request.use` should work', async () => {
-      const xiorInstance = xior.create({ baseURL });
+  describe('`Xior.interceptors` tests', () => {
+    it('`Xior.interceptors.request.use` should work', async () => {
+      const xiorInstance = Xior.create({ baseURL });
       xiorInstance.interceptors.request.use((config) => {
         return merge(config, {
           headers: {
@@ -201,8 +201,8 @@ describe('xior tests', () => {
       assert.strictEqual(data.headers['x-custom-header'], 'x123456');
     });
 
-    it('`xior.interceptors.response.use` should work', async () => {
-      const xiorInstance = xior.create({
+    it('`Xior.interceptors.response.use` should work', async () => {
+      const xiorInstance = Xior.create({
         baseURL,
         headers: { 'x-custom-header': 'x1234567890' },
       });
@@ -219,7 +219,7 @@ describe('xior tests', () => {
 
   describe('should work with basic configs or override configs', () => {
     it('baseURL', async () => {
-      const xiorInstance = xior.create({
+      const xiorInstance = Xior.create({
         baseURL,
       });
       assert.strictEqual(xiorInstance.config?.baseURL, baseURL);
@@ -229,7 +229,7 @@ describe('xior tests', () => {
     });
 
     it('headers', async () => {
-      const xiorInstance = xior.create({
+      const xiorInstance = Xior.create({
         baseURL,
         headers: { 'x-custom-header': 'x1234567890' },
       });
@@ -252,7 +252,7 @@ describe('xior tests', () => {
     });
 
     it('params', async () => {
-      const xiorInstance = xior.create({
+      const xiorInstance = Xior.create({
         baseURL,
         params: { a: 1, b: 2 },
       });
@@ -277,7 +277,7 @@ describe('xior tests', () => {
       body.set('field2', 'val2');
       body.set('file', blob, fileName);
 
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const { data } = await xiorInstance.post<{ file: any; body: Record<string, string> }>(
         '/upload',
         body
@@ -294,7 +294,7 @@ describe('xior tests', () => {
 
   describe('stream test', () => {
     it('stream with `responseType: stream` should work', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const { response } = await xiorInstance.post<{ file: any; body: Record<string, string> }>(
         '/stream/10',
         null,
@@ -309,7 +309,7 @@ describe('xior tests', () => {
     });
 
     it('stream download image should work', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
 
       // GET request for remote image in node.js
       await xiorInstance
@@ -327,7 +327,7 @@ describe('xior tests', () => {
 
   describe('custom AbortController signal should work', () => {
     it('should work with abort early GET', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const controller = new AbortController();
       class CancelFetchError extends Error {}
       setTimeout(() => {
@@ -343,7 +343,7 @@ describe('xior tests', () => {
     });
 
     it('should work with timeout GET', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const controller = new AbortController();
       setTimeout(() => {
         controller.abort(new Error('hi'));
@@ -358,7 +358,7 @@ describe('xior tests', () => {
     });
 
     it('should work with abort early POST', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const controller = new AbortController();
       class CancelFetchError extends Error {}
       setTimeout(() => {
@@ -374,7 +374,7 @@ describe('xior tests', () => {
     });
 
     it('should work with timeout POST', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const controller = new AbortController();
       setTimeout(() => {
         controller.abort(new Error('hi'));
@@ -391,7 +391,7 @@ describe('xior tests', () => {
 
   describe('`responseType` should work', () => {
     it('should work with `responseType: "blob"`', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const { data: getData } = await xiorInstance.get<Blob>('/get', {
         responseType: 'blob',
       });
@@ -406,7 +406,7 @@ describe('xior tests', () => {
     });
 
     it('should work with `responseType: "arraybuffer"`', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const { data: getData } = await xiorInstance.get<ArrayBuffer>('/get', {
         responseType: 'arraybuffer',
       });
@@ -421,7 +421,7 @@ describe('xior tests', () => {
     });
 
     it('should work with as expected `responseType: "original"` or `responseType: "stream"`', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
       const { data: getData } = await xiorInstance.get<ArrayBuffer>('/get', {
         responseType: 'original',
       });
@@ -438,7 +438,7 @@ describe('xior tests', () => {
 
   describe('custom plugins should work', () => {
     it('should work with custom plugin', async () => {
-      const xiorInstance = xior.create({ baseURL });
+      const xiorInstance = Xior.create({ baseURL });
 
       xiorInstance.plugins.use(function plugin1(adapter) {
         return async (config) => {
@@ -471,9 +471,9 @@ describe('xior tests', () => {
   });
 
   describe('interceptors and plugins eject and clear tests', () => {
-    const xiorInstance = xior.create({ baseURL });
+    const xiorInstance = Xior.create({ baseURL });
 
-    it('xior.interceptors.request.use/eject/clear should work', () => {
+    it('Xior.interceptors.request.use/eject/clear should work', () => {
       assert.strictEqual((xiorInstance as any).REQI.length, 0);
       const handler = xiorInstance.interceptors.request.use((config) => {
         return config;
@@ -489,7 +489,7 @@ describe('xior tests', () => {
       assert.strictEqual((xiorInstance as any).REQI.length, 0);
     });
 
-    it('xior.interceptors.response.use/eject/clear should work', () => {
+    it('Xior.interceptors.response.use/eject/clear should work', () => {
       assert.strictEqual((xiorInstance as any).RESI.length, 0);
       const handler = xiorInstance.interceptors.response.use((config) => {
         return config;
@@ -505,7 +505,7 @@ describe('xior tests', () => {
       assert.strictEqual((xiorInstance as any).RESI.length, 0);
     });
 
-    it('xior.plugins.use/eject/clear should work', () => {
+    it('Xior.plugins.use/eject/clear should work', () => {
       assert.strictEqual((xiorInstance as any).P.length, 0);
       const handler = xiorInstance.plugins.use((plugin) => {
         return (config) => {
@@ -527,7 +527,7 @@ describe('xior tests', () => {
   });
 
   describe('custom plugins tests', () => {
-    const xiorInstance = xior.create({ baseURL });
+    const xiorInstance = Xior.create({ baseURL });
     it('xior plugins order should work', async () => {
       assert.strictEqual((xiorInstance as any).P.length, 0);